requirements.txt
selenium==4.4.3
pytest==7.1.3
webdriver-manager==3.8.3
openpyxl==3.0.10
allure-pytest==2.11.0